backend-extension/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
aiofiles==23.2.0
redis==5.0.1